Elevate to sudo and retain permissions during long running scripts
I've added to blocks of code. The first elevates to `sudo` and requests that the user enters their password. The second block ensures that even long-running scripts retain the ability to issue `sudo` after.

I'm not sure that is you include additional scripts or source additional scripts if the `sudo` will be maintained as these will be sub-shells.